(Firebase Authentication - Google Sign In)
The door hissed shut behind Samuel as he stepped into the elvator. He could feel some sense of finality weighing on his shoulders as the elevator started its ascent, but couldn't help but hum to the inane music playing over the speakers. It was an odd comfort in a time where he most certainly needed it.