(Pushing for a miracle)
All of a sudden, the elevator doors began to shudder. Something was causing them to grind against the pistons sealing it shut... The doors slowly started to split open, inhumanly long fingers visible in the breach. They wrap around the elevator, trembling with effort as they pried the door open.