(Please god please god please work)
Seeing those digits coil around the door, Samuel recoiled. Retreating to the farthest corner of the elvator, he could only watch as the doors were split wider and wider. His hand skimmed the floor, searching for something to fight back with, but he only managed to cut himself on the glass shards littering the floor.